+++ kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/config/Defaults.kt
@@ -16,6 +16,8 @@ object Defaults {
    val isolationMode: IsolationMode = IsolationMode.SingleInstance
    val duplicateTestNameMode: DuplicateTestNameMode = DuplicateTestNameMode.Warn
 
+   const val coroutineDebugProbes: Boolean = false
+
    const val failOnEmptyTestSuite = false
 
    const val specFailureFilePath: String = "./.kotest/spec_failures"