+++ kotest-framework/kotest-framework-engine/src/desktopMain/kotlin/io/kotest/engine/withDebugProbe.kt
@@ -1,0 +1,5 @@
+package io.kotest.engine
+
+actual inline fun <T> withDebugProbe(f: () -> T): T {
+   return f()
+}