+++ kotest-framework/kotest-framework-engine/src/jvmMain/kotlin/io/kotest/engine/withDebugProbe.kt
@@ -1,0 +1,18 @@
+package io.kotest.engine
+
+import kotlinx.coroutines.debug.DebugProbes
+
+actual inline fun <T> withDebugProbe(f: () -> T): T {
+   DebugProbes.enableCreationStackTraces = false
+   DebugProbes.sanitizeStackTraces = true
+   return if (!DebugProbes.isInstalled) {
+      DebugProbes.install()
+      try {
+         f()
+      } finally {
+         DebugProbes.uninstall()
+      }
+   } else {
+      f()
+   }
+}